# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BrainBuddy Kids API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterParentBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginParentBody {
  email: string;
  password: string;
}

export interface ParentResponse {
  id: number;
  name: string;
  email: string;
  createdAt: string;
  /** @nullable */
  token?: string | null;
}

export interface CreateChildBody {
  name: string;
  /**
   * @minimum 4
   * @maximum 8
   */
  age: number;
  avatarColor?: string;
}

export interface UpdateChildBody {
  name?: string;
  /**
   * @minimum 4
   * @maximum 8
   */
  age?: number;
  avatarColor?: string;
}

export interface Child {
  id: number;
  parentId: number;
  name: string;
  age: number;
  avatarColor: string;
  createdAt: string;
}

export type CreateCompletionBodyActivityType =
  (typeof CreateCompletionBodyActivityType)[keyof typeof CreateCompletionBodyActivityType];

export const CreateCompletionBodyActivityType = {
  match_shape: "match_shape",
  odd_one_out: "odd_one_out",
  memory_cards: "memory_cards",
  count_objects: "count_objects",
  pattern_builder: "pattern_builder",
  letter_sound: "letter_sound",
  brain_teaser: "brain_teaser",
  reading_words: "reading_words",
} as const;

export type CreateCompletionBodyDifficulty =
  (typeof CreateCompletionBodyDifficulty)[keyof typeof CreateCompletionBodyDifficulty];

export const CreateCompletionBodyDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface CreateCompletionBody {
  activityType: CreateCompletionBodyActivityType;
  /**
   * @minimum 1
   * @maximum 3
   */
  stars: number;
  difficulty: CreateCompletionBodyDifficulty;
  /** @minimum 1 */
  durationSeconds: number;
  correct: boolean;
}

export interface ActivityCompletion {
  id: number;
  childId: number;
  activityType: string;
  stars: number;
  difficulty: string;
  durationSeconds: number;
  correct: boolean;
  completedAt: string;
}

export interface Badge {
  id: number;
  childId: number;
  badgeType: string;
  earnedAt: string;
}

export interface ChildStats {
  totalStars: number;
  totalActivities: number;
  activitiesToday: number;
  starsToday: number;
  timeSpentTodaySeconds: number;
  /** @nullable */
  bestActivity: string | null;
  suggestedNextActivity: string;
  currentDifficulty: string;
  streakCorrect: number;
  totalBadges: number;
}

export interface ActivityBreakdown {
  activityType: string;
  count: number;
  totalStars: number;
  avgStars: number;
}

export interface DailyStar {
  date: string;
  stars: number;
}

export interface ChildProgress {
  recentCompletions: ActivityCompletion[];
  activityBreakdown: ActivityBreakdown[];
  weeklyStars: DailyStar[];
}